moved utils.js file to other folder to make sure unit tests can be executed
